First commit just setting up a blockchain small project.
This commit adds essential building blocks for the chain in block.rs:
1. The block
2. The header

The vision is simple. Let's make blockchain backed identities.

I think it's interesting because similar to money, identities are mostly
backed by things, as countries in real life, and my digital companies on
the web.

However, it doesn't necessarily have to be that way, identities
similarly to money hold value because of a common held belief by the
people using them that they work the way they do.